[GOHW-2]: переход на grpc

+++ checkout/internal/services/cart/add_to_cart.go
@@ -6,13 +6,14 @@ import (
 	"route256/checkout/internal/models"
 )
 
-func (c *Cart) AddToCart(ctx context.Context, user int64, sku uint32, count uint16) error {
+func (c *Cart) AddToCart(ctx context.Context, user int64, sku uint32, count uint32) error {
 	op := "Cart.AddToCart"
 
 	stocks, err := c.lomsClient.Stocks(ctx, sku)
 	if err != nil {
 		return fmt.Errorf("%s: %w", op, err)
 	}
+
 	counter := int64(count)
 	for _, stock := range stocks {
 		counter -= int64(stock.Count)
@@ -23,5 +24,6 @@ func (c *Cart) AddToCart(ctx context.Context, user int64, sku uint32, count uint
 	if counter > 0 {
 		return fmt.Errorf("%s: %w", op, models.ErrInsufficientStocks)
 	}
+
 	return nil
 }
